+++ metricflow/query/query_parser.py
@@ -18,9 +18,9 @@ from metricflow.errors.errors import UnableToSatisfyQueryError
 from metricflow.model.objects.constraints.where import WhereClauseConstraint
 from metricflow.model.objects.elements.dimension import DimensionType
 from metricflow.model.semantic_model import SemanticModel
+from metricflow.model.spec_converters import WhereConstraintConverter
 from metricflow.naming.linkable_spec_name import StructuredLinkableSpecName
 from metricflow.object_utils import pformat_big_objects
-from metricflow.protocols.semantics import DataSourceSemanticsAccessor
 from metricflow.query.query_exceptions import InvalidQueryException
 from metricflow.references import DimensionReference, IdentifierReference, TimeDimensionReference
 from metricflow.specs import (
@@ -65,9 +65,6 @@ class MetricFlowQueryParser:
     "user_id__country" is the "country" dimension that is retrieved by joining "user_id" to the measure data source.
     """
 
-    # TODO: Separate methods out of MetricFlowQueryParser
-    # https://app.asana.com/0/1161293048858925/1201755441255186
-
     # Prefix to use for indents when logging.
     _INDENT_PREFIX = "  "
 
@@ -101,69 +98,6 @@ class MetricFlowQueryParser:
             semantic_model=self._model,
             source_nodes=source_nodes,
             node_output_resolver=node_output_resolver,
-        )
-
-    @staticmethod
-    def convert_to_linkable_specs(
-        data_source_semantics: DataSourceSemanticsAccessor, where_constraint_names: List[str]
-    ) -> LinkableSpecSet:
-        """Processes where_clause_constraint.linkable_names into associated LinkableInstanceSpecs (dims, times, ids)
-
-        where_constraint_names: WhereConstraintClause.linkable_names
-        data_source_semantics: DataSourceSemanticsAccessor from the instantiated class
-
-        output: InstanceSpecSet of Tuple(DimensionSpec), Tuple(TimeDimensionSpec), Tuple(IdentifierSpec)
-        """
-        where_constraint_dimensions = []
-        where_constraint_time_dimensions = []
-        where_constraint_identifiers = []
-        linkable_spec_names = [
-            StructuredLinkableSpecName.from_name(linkable_name) for linkable_name in where_constraint_names
-        ]
-        dimension_references = {
-            dimension_reference.element_name: dimension_reference
-            for dimension_reference in data_source_semantics.get_dimension_references()
-        }
-        identifier_references = {
-            identifier_reference.element_name: identifier_reference
-            for identifier_reference in data_source_semantics.get_identifier_references()
-        }
-
-        for spec_name in linkable_spec_names:
-            if spec_name.element_name == DataSet.metric_time_dimension_name():
-                where_constraint_time_dimensions.append(TimeDimensionSpec.from_name(spec_name.qualified_name))
-            elif spec_name.element_name in dimension_references:
-                dimension = data_source_semantics.get_dimension(dimension_references[spec_name.element_name])
-                if dimension.type == DimensionType.CATEGORICAL:
-                    where_constraint_dimensions.append(DimensionSpec.from_name(spec_name.qualified_name))
-                elif dimension.type == DimensionType.TIME:
-                    where_constraint_time_dimensions.append(TimeDimensionSpec.from_name(spec_name.qualified_name))
-                else:
-                    raise RuntimeError(f"Unhandled type: {dimension.type}")
-            elif spec_name.element_name in identifier_references:
-                where_constraint_identifiers.append(IdentifierSpec.from_name(spec_name.qualified_name))
-            else:
-                raise InvalidQueryException(f"Unknown element: {spec_name}")
-
-        return LinkableSpecSet(
-            dimension_specs=tuple(where_constraint_dimensions),
-            time_dimension_specs=tuple(where_constraint_time_dimensions),
-            identifier_specs=tuple(where_constraint_identifiers),
-        )
-
-    @staticmethod
-    def convert_to_spec_where_constraint(
-        data_source_semantics: DataSourceSemanticsAccessor, where_constraint: WhereClauseConstraint
-    ) -> SpecWhereClauseConstraint:
-        """Converts a where constraint to one using specs."""
-        return SpecWhereClauseConstraint(
-            where_condition=where_constraint.where,
-            linkable_names=tuple(where_constraint.linkable_names),
-            linkable_spec_set=MetricFlowQueryParser.convert_to_linkable_specs(
-                data_source_semantics=data_source_semantics,
-                where_constraint_names=where_constraint.linkable_names,
-            ),
-            execution_parameters=where_constraint.sql_params,
         )
 
     @staticmethod
@@ -414,14 +348,9 @@ class MetricFlowQueryParser:
 
         spec_where_constraint: Optional[SpecWhereClauseConstraint] = None
         if parsed_where_constraint:
-            spec_where_constraint = SpecWhereClauseConstraint(
-                where_condition=parsed_where_constraint.where,
-                linkable_names=tuple(parsed_where_constraint.linkable_names),
-                linkable_spec_set=MetricFlowQueryParser.convert_to_linkable_specs(
-                    data_source_semantics=self._data_source_semantics,
-                    where_constraint_names=parsed_where_constraint.linkable_names,
-                ),
-                execution_parameters=parsed_where_constraint.sql_params,
+            spec_where_constraint = WhereConstraintConverter.convert_to_spec_where_constraint(
+                data_source_semantics=self._data_source_semantics,
+                where_constraint=parsed_where_constraint,
             )
 
         return MetricFlowQuerySpec(
